prospector cleanup - handle cyclic-import
