+++ onadata/apps/api/models/organization_profile.py
@@ -10,10 +10,12 @@ from django.db.models.signals import post_delete, post_save
 
 from guardian.models import GroupObjectPermissionBase, UserObjectPermissionBase
 from guardian.shortcuts import assign_perm, get_perms_for_model
+from multidb.pinning import use_master
 
 from onadata.apps.api.models.team import Team
-from onadata.apps.main.models import UserProfile
+from onadata.apps.main.models.user_profile import UserProfile
 from onadata.libs.utils.cache_tools import IS_ORG, safe_delete
+from onadata.libs.utils.common_tags import MEMBERS
 
 User = get_user_model()
 
@@ -62,6 +64,82 @@ def create_owner_team_and_assign_permissions(org):
 
             if org.created_by and org.created_by != org.creator:
                 assign_perm(perm.codename, org.created_by, org.userprofile_ptr)
+
+    return team
+
+
+# pylint: disable=invalid-name
+def get_or_create_organization_owners_team(org):
+    """
+    Get the owners team of an organization
+    :param org: organization
+    :return: Owners team of the organization
+    """
+    team_name = f"{org.user.username}#{Team.OWNER_TEAM_NAME}"
+    try:
+        team = Team.objects.get(name=team_name, organization=org.user)
+    except Team.DoesNotExist:
+        with use_master:
+            queryset = Team.objects.filter(name=team_name, organization=org.user)
+            if queryset.count() > 0:
+                return queryset.first()  # pylint: disable=no-member
+            return create_owner_team_and_assign_permissions(org)
+    return team
+
+
+def add_user_to_team(team, user):
+    """
+    Adds a user to a team and assigns them team permissions.
+    """
+    user.groups.add(team)
+
+    # give the user perms to view the team
+    assign_perm("view_team", user, team)
+
+    # if team is owners team assign more perms
+    if team.name.find(Team.OWNER_TEAM_NAME) > 0:
+        _assign_organization_team_perms(team.organization, user)
+
+
+def _assign_organization_team_perms(organization, user):
+    owners_team = get_or_create_organization_owners_team(organization.profile)
+    members_team = get_organization_members_team(organization.profile)
+    for perm in get_perms_for_model(Team):
+        assign_perm(perm.codename, user, owners_team)
+        assign_perm(perm.codename, user, members_team)
+
+
+def create_organization_team(organization, name, permission_names=None):
+    """
+    Creates an organization team with the given permissions as defined in
+    permission_names.
+    """
+    organization = (
+        organization.user
+        if isinstance(organization, OrganizationProfile)
+        else organization
+    )
+    team = Team.objects.create(organization=organization, name=name)
+    content_type = ContentType.objects.get(app_label="api", model="organizationprofile")
+    if permission_names:
+        # get permission objects
+        perms = Permission.objects.filter(
+            codename__in=permission_names, content_type=content_type
+        )
+        if perms:
+            team.permissions.add(*tuple(perms))
+    return team
+
+
+def get_organization_members_team(organization):
+    """Get organization members team
+    create members team if it does not exist and add organization owner
+    to the members team"""
+    try:
+        team = Team.objects.get(name=f"{organization.user.username}#{MEMBERS}")
+    except Team.DoesNotExist:
+        team = create_organization_team(organization, MEMBERS)
+        add_user_to_team(team, organization.user)
 
     return team
