+++ onadata/apps/api/tools.py
@@ -1,6 +1,6 @@
 # -*- coding=utf-8 -*-
 """
-API util functions.
+API utility functions.
 """
 import os
 import tempfile
@@ -9,8 +9,6 @@ from datetime import datetime
 from django import forms
 from django.conf import settings
 from django.contrib.auth import get_user_model
-from django.contrib.auth.models import Permission
-from django.contrib.contenttypes.models import ContentType
 from django.contrib.sites.models import Site
 from django.core.files.storage import get_storage_class
 from django.core.files.uploadedfile import InMemoryUploadedFile
@@ -23,7 +21,7 @@ from django.shortcuts import get_object_or_404
 from django.utils.module_loading import import_string
 from django.utils.translation import gettext as _
 
-from guardian.shortcuts import assign_perm, get_perms, get_perms_for_model, remove_perm
+from guardian.shortcuts import get_perms, get_perms_for_model, remove_perm
 from kombu.exceptions import OperationalError
 from multidb.pinning import use_master
 from registration.models import RegistrationProfile
@@ -33,7 +31,9 @@ from taggit.forms import TagField
 
 from onadata.apps.api.models.organization_profile import (
     OrganizationProfile,
-    create_owner_team_and_assign_permissions,
+    add_user_to_team,
+    get_or_create_organization_owners_team,
+    get_organization_members_team,
 )
 from onadata.apps.api.models.team import Team
 from onadata.apps.logger.models import DataView, Instance, Project, XForm
@@ -183,60 +183,6 @@ def create_organization_object(org_name, creator, attrs=None):
     return profile
 
 
-def create_organization_team(organization, name, permission_names=None):
-    """
-    Creates an organization team with the given permissions as defined in
-    permission_names.
-    """
-    organization = (
-        organization.user
-        if isinstance(organization, OrganizationProfile)
-        else organization
-    )
-    team = Team.objects.create(organization=organization, name=name)
-    content_type = ContentType.objects.get(app_label="api", model="organizationprofile")
-    if permission_names:
-        # get permission objects
-        perms = Permission.objects.filter(
-            codename__in=permission_names, content_type=content_type
-        )
-        if perms:
-            team.permissions.add(*tuple(perms))
-    return team
-
-
-def get_organization_members_team(organization):
-    """Get organization members team
-    create members team if it does not exist and add organization owner
-    to the members team"""
-    try:
-        team = Team.objects.get(name=f"{organization.user.username}#{MEMBERS}")
-    except Team.DoesNotExist:
-        team = create_organization_team(organization, MEMBERS)
-        add_user_to_team(team, organization.user)
-
-    return team
-
-
-# pylint: disable=invalid-name
-def get_or_create_organization_owners_team(org):
-    """
-    Get the owners team of an organization
-    :param org: organization
-    :return: Owners team of the organization
-    """
-    team_name = f"{org.user.username}#{Team.OWNER_TEAM_NAME}"
-    try:
-        team = Team.objects.get(name=team_name, organization=org.user)
-    except Team.DoesNotExist:
-        with use_master:
-            queryset = Team.objects.filter(name=team_name, organization=org.user)
-            if queryset.count() > 0:
-                return queryset.first()  # pylint: disable=no-member
-            return create_owner_team_and_assign_permissions(org)
-    return team
-
-
 def remove_user_from_organization(organization, user):
     """Remove a user from an organization"""
     team = get_organization_members_team(organization)
@@ -281,28 +227,6 @@ def add_user_to_organization(organization, user):
 
     team = get_organization_members_team(organization)
     add_user_to_team(team, user)
-
-
-def add_user_to_team(team, user):
-    """
-    Adds a user to a team and assigns them team permissions.
-    """
-    user.groups.add(team)
-
-    # give the user perms to view the team
-    assign_perm("view_team", user, team)
-
-    # if team is owners team assign more perms
-    if team.name.find(Team.OWNER_TEAM_NAME) > 0:
-        _assign_organization_team_perms(team.organization, user)
-
-
-def _assign_organization_team_perms(organization, user):
-    owners_team = get_or_create_organization_owners_team(organization.profile)
-    members_team = get_organization_members_team(organization.profile)
-    for perm in get_perms_for_model(Team):
-        assign_perm(perm.codename, user, owners_team)
-        assign_perm(perm.codename, user, members_team)
 
 
 def get_organization_members(organization):
