+++ onadata/apps/logger/models/__init__.py
@@ -10,9 +10,9 @@ from onadata.apps.logger.models.note import Note  # noqa
 from onadata.apps.logger.models.open_data import OpenData  # noqa
 from onadata.apps.logger.models.osmdata import OsmData  # noqa
 from onadata.apps.logger.models.project import Project  # noqa
+from onadata.apps.logger.models.submission_review import SubmissionReview  # noqa
 from onadata.apps.logger.models.survey_type import SurveyType  # noqa
 from onadata.apps.logger.models.widget import Widget  # noqa
 from onadata.apps.logger.models.xform import XForm  # noqa
-from onadata.apps.logger.models.submission_review import SubmissionReview  # noqa
-from onadata.apps.logger.xform_instance_parser import InstanceParseError  # noqa
 from onadata.apps.logger.models.xform_version import XFormVersion  # noqa
+from onadata.apps.logger.xform_instance_parser import InstanceParseError  # noqa
