+++ onadata/libs/utils/project_utils.py
@@ -9,7 +9,8 @@ from django.db import IntegrityError
 
 from multidb.pinning import use_master
 
-from onadata.apps.logger.models import Project, XForm
+from onadata.apps.logger.models.project import Project
+from onadata.apps.logger.models.xform import XForm
 from onadata.celery import app
 from onadata.libs.permissions import (
     ROLES,
